WIP
This is a set of fixes to deal with some problems where renovate can't find an
expected symbolic target for a jump.

The first problem was that renovate and macaw had different ideas of what a
block terminators looks like, which caused renovate to look for blocks where
there were none.  The particular trigger seemed to be that, in PowerPC, `b +4`
is identical to a noop to macaw, but renovate thought it was a block terminator.
The fix to that was to just trust macaw and use the block lengths that it gives
us.

The current problem is that some blocks appear multiple times with different
lengths in the results from macaw, which renovate doesn't expect.  It isn't
exactly obvious what to do in these cases.  We might have to make it so that
renovate is able to deal with overlapping blocks robustly and just have duplicates.