WIP
Add some code to detect overlapping blocks and avoid ever processing them.

The current problem for PowerPC is that the code discovery is considering TOC
data as code, which produces lots of bad false positives.  Unfortunately, this
memory really is mapped (and executable).

To fix this, we need to pass in the range of addresses we actually want to
consider valid code (taken from the boundaries of the text section).  We have
this available.