Restrict code recovery to the address range of the text section
One of the current problems is that macaw aggressively explores all addresses
written to memory as code, but this picks up the TOC in PowerPC, which happens
to be mapped executable.  Treating this as code causes all sorts of issues.

Fixing that reveals the next problem: there are some legitimate jumps outside of
the text section (e.g., to .init or .fini).  Since we aren't exploring them
anymore, symbolization fails because some jump targets are missing.  To fix
that, we need to make symbolization robust to missing targets (which is fine --
we can just change how we fix things up during layout).