WIP Extend SymbolicAddress
This commit updates the renovate core with an extended symbolic address type.
The problem is that the old symbolic addressing method was unable to handle
references to addresses outside of the text section (which is a big problem in
practice).  Now, there is a notion of a reference to an externally stable
address (from an instruction address that might move).

The core is updated, but the arch packages are not.